// lib/api/generated/models/VlogMediaOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VlogMediaOut = {
    media_type: 'concept-image' | 'concept-video' | 'chapter-cover' | 'chapter-trailer' | 'episode-video' | 'episode-thumbnail' | 'short-film';
    fileId?: (string | null);
    name?: (string | null);
    url: string;
    thumbnailUrl?: (string | null);
    filePath?: (string | null);
    mime?: (string | null);
    size?: (number | null);
    id: string;
    created_at: string;
};